Feature/registration

+++ src/main/java/edu/ftn/iss/eventplanner/repositories/ServiceAndProductProviderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ServiceAndProductProviderRepository extends JpaRepository<ServiceAndProductProvider, Long> {
+public interface ServiceAndProductProviderRepository extends JpaRepository<ServiceAndProductProvider, Integer> {
     Optional<ServiceAndProductProvider> findByEmail(String email);
     ServiceAndProductProvider findByActivationToken(String token);
 }
